refactor: move queryParamToStr func to seperate file
