+++ api/controllers/OpenEAppController.js
@@ -63,11 +63,10 @@ const OpenEAppController = {
 
     async _getApplicationDataFromCasebook(req, res) {
         try {
-            const queryParamsObj = {
-                timestamp: new Date().getTime().toString(),
-                applicationReference: req.params.unique_app_id,
-            };
-            return await CasebookService.getApplicationStatus(queryParamsObj);
+            const applicationReference = req.params.unique_app_id;
+            return await CasebookService.getApplicationStatus(
+                applicationReference
+            );
         } catch (error) {
             sails.log.error(error);
             return res.serverError();
